+++ CHANGELOG.md
@@ -37,7 +37,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Detect OIDC tokens issued by Kubernetes (smallstep/cli#953).
 - Add support for Smallstep Managed Endpoint X509 extension
   (smallstep/cli#989).
-- Support signing a certificate for a private key that can only be used for encryption with the `--skip-csr-signature` flag in `step certificate create`. Some KMSs restrict key usage to a single type of cryptographic operation. This blocks RSA decryption keys from being used to sign a CSR for their public key. Using the `--skip-csr-signature` flag, the public key is used directly with a certificate template, removing the need for the CSR signature.
+- Support signing a certificate for a private key that can only be used for 
+  encryption with the `--skip-csr-signature` flag in `step certificate create`. 
+  Some KMSs restrict key usage to a single type of cryptographic operation. 
+  This blocks RSA decryption keys from being used to sign a CSR for their public 
+  key. Using the `--skip-csr-signature` flag, the public key is used directly 
+  with a certificate template, removing the need for the CSR signature.
+- Add all AWS identity document certificates (smallstep/certificates#1510)
 
 ### Changed
 
@@ -50,6 +56,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   (smallstep/certificates#976)
 - Fix empty ca.json with invalid template data (smallstep/certificates#1501).
 - Fix interactive prompt on docker builds (smallstep/cli#963).
+- `step certificate fingerprint` correctly parse PEM files with non-PEM header
+  (smallstep/crypto#311)
+- `step certificate format` correctly parse PEM files with non-PEM header
+  (smallstep/cli#1006)
 
 ### Remove